Added Configuration

+++ MinecraftMediaLibrary/src/main/java/com/github/pulsebeat02/image/MapImage.java
@@ -37,7 +37,7 @@ public class MapImage implements AbstractImageMapHolder, ConfigurationSerializab
     }
 
     public MapImage(@NotNull final MinecraftMediaLibrary library,
-                    final int map,
+                    final long map,
                     @NotNull final String url,
                     final int width,
                     final int height) {
